Decomposed YoubotJoint.hpp
